Android crystax model fitting and pose separation:
Model fitting is working on android crystax. Had to change some cvRound functions temporarily and disable the black line fix for the isomap. Will revert these changes when there is a fix for the crystax bug.

Also split pose estimation from model fitting and added a data structure to store calculations done by the pose estimation that can be used later to finish fitting. This is handy if you want to collect a series of images from the camera based on subject pose, then do the model fitting in a single step later.